requirements.txt
# Core dependencies
python-dotenv>=1.0.0
langchain-openai>=0.3.0
langgraph>=0.2.0
pydantic>=2.0.0

# Databricks SDK
databricks-sdk>=0.40.0
